cleanup marioAccess

+++ src/Game/Player/MarioActorMorph.cpp
@@ -16,7 +16,7 @@
 #include "revolution/types.h"
 
 void MarioActor::setPlayerMode(u16 playerMode, bool myBool) {
-    if (mPlayerMode == (u16)playerMode) {
+    if (mPlayerMode == playerMode) {
         if (playerMode == 1) {
             _3DC = mConst->getTable()->mMetalHoldTime;
             MR::startSubBGM("BGM_MUTEKI_A", false);
